unify ip_* and ip6_*
Just like sources_* and sources6_*, it would be nice if you didn't have to
worry about whether addresses on your interfaces are IPv4 or IPv6.
This change effects that (backwards compatible, of course).

As a side effect this introduces the somewhat peculiar (but useful) feature
that you can assign multiple addresses to a {interface}_{name} that will
all be treated the same.

In effect the name no longer describes a single address but a class of
addresses that all get the same rules. This can greatly reduce duplication
in configurations with large numbers of addresses (something not unheard of
with IPv6).

This patch also makes some formatting/style choices more consistent/uniform:
 * use case statements wherever possible (to reduce quoting issues);
 * deobfuscate eval statements by removing redundant escaped quotes;
 * always use short option syntax (when possible) for single-line iptables
   invocations;
 * always use long option syntax for iptables invocations that are spread
   out over multiple lines.

Signed-off-by: Joost van Baal-Ilić <joostvb@uvt.nl>