unify net_* and net6_*
Unifies the net_* and net6_ variables just like sources_* and sources6_*
(and ip_* and ip6_*).

As a nice side-effect this finally allows multiple networks to be specified
in net_* (and net6_*).

Also expanded the list of bogon networks, based on RFC 6890:

	0.0.0.0/8
	10.0.0.0/8
	100.64.0.0/10
	127.0.0.0/8
	169.254.0.0/16
	172.16.0.0/12
	192.0.0.0/24
	192.0.2.0/24
	192.88.99.0/24
	192.168.0.0/16
	198.18.0.0/15
	198.51.100.0/24
	203.0.113.0/24
	224.0.0.0/3

	64:ff9b::/96
	::ffff:0:0/96
	100::/64
	200::/7
	2001:2::/48
	2001:db8::/32
	2001:10::/28
	fc00::/7
	fec0::/10
	3ffe::/16
	5f00::/8
	::1/128
	::/128

Improved uruk_save() option snooping to accept other syntax variants
(-tnat, --table=nat).

Made the syntax of uruk-save consistent with that of the main uruk script.

Signed-off-by: Joost van Baal-Ilić <joostvb@uvt.nl>